+++ aioplus/__init__.py
@@ -2,6 +2,7 @@ from aioplus.internal.aall import aall
 from aioplus.internal.aany import aany
 from aioplus.internal.abatched import abatched
 from aioplus.internal.acount import acount
+from aioplus.internal.acycle import acycle
 from aioplus.internal.aenumerate import aenumerate
 from aioplus.internal.aislice import aislice
 from aioplus.internal.alen import alen
@@ -13,13 +14,14 @@ from aioplus.internal.awaitify import awaitify
 
 
 __author__ = "Sergei Y. Bogdanov <syubogdanov@outlook.com>"
-__version__ = "0.0.0"
+__version__ = "0.1.0"
 
 __all__: list[str] = [
     "aall",
     "aany",
     "abatched",
     "acount",
+    "acycle",
     "aenumerate",
     "aislice",
     "alen",
@@ -35,6 +37,7 @@ aall.__module__ = "aioplus"
 aany.__module__ = "aioplus"
 abatched.__module__ = "aioplus"
 acount.__module__ = "aioplus"
+acycle.__module__ = "aioplus"
 aenumerate.__module__ = "aioplus"
 aislice.__module__ = "aioplus"
 alen.__module__ = "aioplus"