+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Added `aioplus.acycle`;
 * Added `aioplus.apairwise`.
 
 ## [0.0.0] - 2025-07-27